requirements.txt
fastapi
uvicorn
google-cloud-vision
python-multipart
pydantic